release: rewrite NavbarItem to use typescript, and to function on preview sites
